[#25] Apply template-method pattern to interactors
